pyproject.toml
[project]
name = "binny"
version = "0.1.0"
description = "CLI inventory management chatbot with part naming capabilities"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "claude-agent-sdk>=0.1.4",
    "rich>=14.2.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
binny = "binny.main:main"
